chore: remove Paper view managers, add defensive JSON parsing
