+++ android/src/main/java/com/adyenreactnativesdk/react/PlatformPayViewManager.kt
@@ -1,5 +1,6 @@
 package com.adyenreactnativesdk.react
 
+import com.adyenreactnativesdk.react.base.OnPressEvent
 import com.adyenreactnativesdk.react.platformpay.ButtonTheme
 import com.adyenreactnativesdk.react.platformpay.ButtonType
 import com.adyenreactnativesdk.react.platformpay.PlatformPayView
@@ -11,7 +12,6 @@ import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.uimanager.UIManagerHelper
 import com.facebook.react.uimanager.ViewManagerDelegate
 import com.facebook.react.uimanager.annotations.ReactProp
-import com.facebook.react.uimanager.events.Event
 import com.facebook.react.viewmanagers.PlatformPayViewManagerDelegate
 import com.facebook.react.viewmanagers.PlatformPayViewManagerInterface
 import com.google.android.gms.wallet.button.ButtonConstants
@@ -74,19 +74,5 @@ class PlatformPayViewManager :
     val eventDispatcher = UIManagerHelper.getEventDispatcherForReactTag(context, viewId)
     val event = OnPressEvent(surfaceId, viewId)
     eventDispatcher?.dispatchEvent(event)
-  }
-
-  override fun getExportedCustomDirectEventTypeConstants(): Map<String, Any> =
-    mapOf(OnPressEvent.EVENT_NAME to mapOf("registrationName" to "onButtonPress"))
-}
-
-class OnPressEvent(
-  surfaceId: Int,
-  viewId: Int,
-) : Event<OnPressEvent>(surfaceId, viewId) {
-  override fun getEventName() = EVENT_NAME
-
-  companion object {
-    const val EVENT_NAME: String = "onButtonPress"
   }
 }
