+++ android/src/main/java/com/adyenreactnativesdk/react/base/LayoutChangeEvent.kt
@@ -4,12 +4,12 @@ import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.uimanager.events.Event
 
-class ResizableCustomViewEvent(
+class LayoutChangeEvent(
   surfaceId: Int,
   viewId: Int,
   private val width: Int,
   private val height: Int,
-) : Event<ResizableCustomViewEvent>(surfaceId, viewId) {
+) : Event<LayoutChangeEvent>(surfaceId, viewId) {
   override fun getEventName() = EVENT_NAME
 
   override fun getEventData(): WritableMap =
