--- ios/Views/CardView/ADYCardViewManager.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#import <React/RCTViewManager.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface ADYCardViewManager : RCTViewManager
-@end
-
-NS_ASSUME_NONNULL_END
