--- ios/Views/CardView/ADYCardViewManager.mm
@@ -1,22 +1,0 @@
-#import "ADYCardView.h"
-#import "RCTBridge.h"
-#import <React/RCTUIManager.h>
-#import <React/RCTViewManager.h>
-#include <objc/objc.h>
-
-@interface ADYCardViewManager : RCTViewManager
-@end
-
-@implementation ADYCardViewManager
-
-RCT_EXPORT_MODULE(CardView)
-
-- (UIView *)view {
-  return [[ADYCardView alloc] init];
-}
-
-RCT_EXPORT_VIEW_PROPERTY(paymentMethod, NSString)
-RCT_EXPORT_VIEW_PROPERTY(configuration, NSString)
-RCT_EXPORT_VIEW_PROPERTY(onLayoutChange, RCTDirectEventBlock)
-
-@end
