--- ios/Views/PlatformView/ADYPlatformPayViewManager.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#import <React/RCTViewManager.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface ADYPlatformPayViewManager : RCTViewManager
-
-@end
-
-NS_ASSUME_NONNULL_END
